Change int to double for better accuracy in calculating. When we use int and rotate some element he collapses

+++ GTLib/Primitives/Dot2D.cs
@@ -7,10 +7,10 @@ namespace GTLib.Primitives
 {
     public class Dot2D: Primitive2D
     {
-        public int X { get; set; }
-        public int Y { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
 
-        public Dot2D(int X,int Y)
+        public Dot2D(double X, double Y)
         {
             this.X = X;
             this.Y = Y;
